requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.4
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.137
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==3.20.3
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.49
streamlit==1.18.1
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.20.0
yarl==1.9.2
zipp==3.16.2
